Fixed the bugs introduced from last commit
Finally got motivated (for once) to finnish these bad boys. Ready for merge


Former-commit-id: 64a03dcd009cd54a154a4c27b42e6778b98f1261

+++ Intersect.Server/Classes/Entities/Player.cs
@@ -1691,9 +1691,9 @@ namespace Intersect.Server.Entities
             return lifesteal;
         }
 
-        public int GetTenacity()
+        public double GetTenacity()
         {
-          int tenacity = 0;
+          double tenacity = 0;
 
           for (var i = 0; i < Options.EquipmentSlots.Count; i++)
           {
@@ -1716,9 +1716,9 @@ namespace Intersect.Server.Entities
           return tenacity;
         }
 
-        public int GetLuck()
+        public double GetLuck()
         {
-          int luck = 0;
+          double luck = 0;
 
           for (var i = 0; i < Options.EquipmentSlots.Count; i++)
           {
